+++ modules/light-clients/08-wasm/client/cli/tx.go
@@ -1,6 +1,7 @@
 package cli
 
 import (
+	"encoding/hex"
 	"fmt"
 	"os"
 
@@ -82,3 +83,49 @@ func newSubmitStoreCodeProposalCmd() *cobra.Command {
 
 	return cmd
 }
+<<<<<<< HEAD
+=======
+
+func newMigrateContractCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:     "migrate-contract [client-id] [checksum] [migrate-msg]",
+		Short:   "Migrates a contract to a new byte code",
+		Long:    "Migrates the contract for the specified client ID to the byte code corresponding to checksum, passing the JSON-encoded migrate message to the contract",
+		Example: fmt.Sprintf("%s tx %s-wasm migrate-contract 08-wasm-0 b3a49b2914f5e6a673215e74325c1d153bb6776e079774e52c5b7e674d9ad3ab {}", version.AppName, ibcexported.ModuleName),
+		Args:    cobra.ExactArgs(3), // Ensure exactly three arguments are passed
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			clientID := args[0]
+			checksumHex := args[1]
+			migrateMsg := args[2]
+
+			checksum, err := hex.DecodeString(checksumHex)
+			if err != nil {
+				return fmt.Errorf("invalid checksum format: %w", err)
+			}
+
+			// Construct the message
+			msg := &types.MsgMigrateContract{
+				Signer:   clientCtx.GetFromAddress().String(),
+				ClientId: clientID,
+				Checksum: checksum,
+				Msg:      []byte(migrateMsg),
+			}
+
+			if err := msg.ValidateBasic(); err != nil {
+				return err
+			}
+
+			// Generate or broadcast the transaction
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+	return cmd
+}
+>>>>>>> d51ab023 (fix(08-wasm): decode checksum hex string to bytes in migrate contract CLI (#6815))